move swerve module and motor code to own folder
