+++ src/main/java/frc/robot/subsystems/swerve/DriveMotor.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.swerve;
 
 
 import com.ctre.phoenix6.configs.CurrentLimitsConfigs;
