+++ src/main/java/frc/robot/subsystems/swerve/SwerveModule.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.swerve;
 
 
 import java.text.DecimalFormat;
